Change project direction
The design is much simpler and avoids having a complex query builder to
memorize each element of the SQL statement to generate. There is now a
single method call with a single block (instead of a block per
statement) to generate a single NamedTuple for each query.

It also solves how to alias table names and columns.

It requires the schemas to be explicit, but I think this is a much
better approach overall (it will eventually be auto-generated from
database introspection).