Change project direction (again)
The revamped design using a NamedTuple was pretty nice, I especially
liked how it read, but there was some issues: the design was very rigid,
which led the implementation to be hard to extend, and I couldn't get
sub-queries to be implemented correctly.

This redesign returns to chained methods, but immediately builds the SQL
string and args array, instead of trying to memorize everything into a
data structure. Each query builder method is thus as dumb as possible,
and limited at generating SQL as you'd write it manually.

There are still issues with sub-queries, but using blocks to be invoked
instead of memoizing SQL::Builder instances should solve them.